requirements.txt
torch==1.13.1
torchvision==0.14.1
transformers==4.21.3
huggingface_hub==0.10.1
sentence-transformers==2.2.2
langchain==0.1.14
langchain-community==0.0.38
langchain-pinecone==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
flask==3.1.1
pypdf==5.6.1
python-dotenv==1.1.0
pydantic==1.10.12
pinecone-client>=3.2.2,<4.0.0
-e .